requirements-macos.txt
# macOS-Compatible Requirements for LANISTR
# This file contains all dependencies needed for local development on macOS

# =============================================================================
# CORE MACHINE LEARNING FRAMEWORKS
# =============================================================================

# PyTorch ecosystem (macOS compatible)
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
torchaudio>=2.0.0,<3.0.0
torchmetrics>=0.9.3,<1.0.0

# Transformers for NLP models
transformers>=4.26.0,<5.0.0

# =============================================================================
# DATA PROCESSING AND ANALYSIS
# =============================================================================

# Core data science libraries
numpy>=1.21.0,<2.0.0
pandas>=1.3.5,<2.0.0
scikit-learn>=1.3.2,<2.0.0

# Image processing
Pillow>=8.0.0,<10.0.0

# =============================================================================
# CONFIGURATION AND UTILITIES
# =============================================================================

# Configuration management
omegaconf>=2.3.0,<3.0.0

# Progress bars and utilities
tqdm>=4.62.0,<5.0.0
pytz>=2021.3,<2024.0

# =============================================================================
# GOOGLE CLOUD PLATFORM INTEGRATION
# =============================================================================

# Core Google Cloud services
google-cloud-storage>=2.0.0,<3.0.0
google-cloud-aiplatform>=1.25.0,<2.0.0
google-auth>=2.0.0,<3.0.0
google-auth-oauthlib>=0.4.0,<2.0.0
google-auth-httplib2>=0.1.0,<1.0.0

# Additional GCP services for enterprise features
google-cloud-logging>=3.0.0,<4.0.0
google-cloud-monitoring>=2.0.0,<3.0.0
google-cloud-error-reporting>=1.0.0,<2.0.0
google-cloud-trace>=1.0.0,<2.0.0

# =============================================================================
# ENTERPRISE FEATURES AND MONITORING
# =============================================================================

# Logging and monitoring
structlog>=21.0.0,<24.0.0
prometheus-client>=0.12.0,<1.0.0

# Health checks and diagnostics
psutil>=5.8.0,<6.0.0
GPUtil>=1.4.0,<2.0.0

# Configuration validation
pydantic>=1.8.0,<2.0.0
jsonschema>=3.2.0,<4.0.0

# =============================================================================
# SECURITY AND COMPLIANCE
# =============================================================================

# Security scanning and validation
bandit>=1.7.0,<2.0.0
safety>=1.10.0,<2.0.0

# Encryption and secure communication
cryptography>=3.4.0,<4.0.0
certifi>=2021.0.0,<2024.0.0

# =============================================================================
# TESTING AND QUALITY ASSURANCE
# =============================================================================

# Testing frameworks
pytest>=6.0.0,<8.0.0
pytest-cov>=2.12.0,<4.0.0
pytest-mock>=3.6.0,<4.0.0

# Code quality and linting
black>=21.0.0,<24.0.0
flake8>=3.9.0,<7.0.0
isort>=5.9.0,<6.0.0
mypy>=0.910,<2.0.0

# =============================================================================
# PERFORMANCE AND OPTIMIZATION
# =============================================================================

# Memory and performance monitoring
memory-profiler>=0.60.0,<1.0.0
line-profiler>=3.3.0,<4.0.0

# =============================================================================
# DATA VALIDATION AND SANITIZATION
# =============================================================================

# Data validation
great-expectations>=0.15.0,<1.0.0
pandera>=0.7.0,<1.0.0

# =============================================================================
# MODEL SERIALIZATION AND DEPLOYMENT
# =============================================================================

# Model serving and serialization
torchserve>=0.6.0,<1.0.0
mlflow>=1.20.0,<2.0.0

# =============================================================================
# NETWORKING AND HTTP
# =============================================================================

# HTTP client and server
requests>=2.25.0,<3.0.0
urllib3>=1.26.0,<2.0.0

# =============================================================================
# UTILITIES AND HELPERS
# =============================================================================

# Date and time utilities
python-dateutil>=2.8.0,<3.0.0

# =============================================================================
# DEVELOPMENT AND DEBUGGING
# =============================================================================

# Debugging tools (optional for production)
ipython>=7.0.0,<9.0.0
jupyter>=1.0.0,<2.0.0

# =============================================================================
# ENTERPRISE INTEGRATION
# =============================================================================

# API documentation
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0

# Database connectors (if needed)
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0

# =============================================================================
# DEPLOYMENT NOTES
# =============================================================================

# For local development on macOS:
# 1. Use a virtual environment
# 2. Install with: pip install -r requirements-macos.txt
# 3. For production deployment, use requirements_vertex_ai.txt on Linux with CUDA 